+++ roco/config.py
@@ -11,6 +11,7 @@ class Settings(BaseSettings):
     oidc_authorize_url: str | None = None
     oidc_redirect_url: str | None = None
     oidc_logout_url: str | None = None
+    oidc_post_logout_redirect_url: str | None = None
     oidc_scope: str = 'openid email'
 
     login_provider: Literal['db', 'ldap'] = 'db'
@@ -24,25 +25,6 @@ class Settings(BaseSettings):
             return 'ldap'
 
         return 'db'
-
-    @model_validator(mode='after')
-    def check_oidc_params(self):
-        three_values = [
-            self.oidc_client_id,
-            self.oidc_authorize_url,
-            self.oidc_redirect_url,
-            self.oidc_logout_url
-        ]
-        count = len([v for v in three_values if v])
-
-        if count not in (0, 4):
-            raise ValueError(
-                'oidc_client_id, oidc_authorize_url, oidc_redirect_url'
-                ' and oidc_logout_url'
-                ' should be either all absent or all present'
-            )
-
-        return self
 
 
 def get_settings():